add breadcrumb molecule, cleanup newplugin files
