+++ web/src/components/molecules/UpdatePluginPage/index.tsx
@@ -1,6 +1,6 @@
 import { UploadRequestOption } from "rc-upload/lib/interface";
 
-import { FileUploadType } from "@marketplace/components/molecules/AddNewPluginPage/PackageArea";
+import { FileUploadType } from "@marketplace/components/molecules/Common/PluginEditing/PackageArea";
 import { styled } from "@marketplace/theme";
 
 import UpdatePluginContent from "./UpdatePluginContent";
